infra/src/requirements.txt
functions-framework==3.3.0
google-cloud-bigquery
google-cloud-storage
